requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.0.1
python-decouple==3.8
beautifulsoup4==4.12.2
requests==2.31.0
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9